requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
attrs==25.3.0
av==14.4.0
blinker==1.9.0
bpy==4.5.2
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
comm==0.2.3
commonmark==0.9.1
configargparse==1.7.1
contourpy==1.3.3
cryptography==45.0.6
cycler==0.12.1
cython==3.1.3
dash==3.2.0
dash-bootstrap-components==2.0.3
dash-canvas==0.1.0
datasets==4.0.0
decorator==5.2.1
dill==0.3.8
dnspython==2.7.0
docutils==0.22
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
filelock==3.19.1
flask==3.1.1
fonttools==4.59.1
freetype-py==2.5.1
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
google-crc32c==1.7.1
h11==0.16.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
ifaddr==0.2.0
imageio==2.37.0
importlib-metadata==6.11.0
inquirerpy==0.3.4
ipython==9.4.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-core==5.8.1
jupyterlab-widgets==3.0.15
kiwisolver==1.4.9
lazy-loader==0.4
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.1.2
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.2
open3d==0.19.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
parso==0.8.4
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.8
plotly==6.3.0
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==4.25.8
pure-eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pyee==13.0.0
pyglet==2.1.6
pygments==2.19.2
pylibsrtp==0.12.0
pyopengl==3.1.0
pyopenssl==25.1.0
pyparsing==3.2.3
pyqt5==5.15.11
pyqt5-qt5==5.15.2
pyqt5-sip==12.17.0
pyrealsense2==2.56.5.9235
pyrender==0.1.45
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
pywin32==311
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
retrying==1.4.2
rich==12.6.0
rich-cli==1.8.1
rich-rst==1.3.1
rpds-py==0.27.0
safetensors==0.6.2
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.47.3
streamlit==1.29.0
streamlit-drawable-canvas==0.9.3
streamlit-image-coordinates==0.4.0
streamlit-webrtc==0.63.4
sympy==1.14.0
tenacity==8.5.0
textual==0.1.18
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
trimesh==4.7.3
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
validators==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
werkzeug==3.1.3
widgetsnbextension==4.0.14
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.24.0
